+++ sbuildr/project/dependencies/builders/sbuildr_builder.py
@@ -1,4 +1,6 @@
 from sbuildr.project.dependencies.builders.builder import DependencyBuilder
+from sbuildr.project.dependencies.meta import DependencyMetadata
+from sbuildr.graph.node import Library
 from sbuildr.project.project import Project
 from sbuildr.misc import utils
 from sbuildr.logger import G_LOGGER
@@ -7,17 +9,19 @@ import sys
 import os
 
 class SBuildrBuilder(DependencyBuilder):
-    def __init__(self, build_script_path: str="build.py", project_save_path: str=Project.DEFAULT_SAVED_PROJECT_NAME):
+    def __init__(self, build_script_path: str="build.py", project_save_path: str=Project.DEFAULT_SAVED_PROJECT_NAME, install_profile="release"):
         f"""
         Builds projects using the SBuildr build system.
 
         :param build_script_path: The path to the build script, relative to the project root. Defaults to "build.py".
         :param project_save_path: The path at which the build script saves the project. Defaults to {Project.DEFAULT_SAVED_PROJECT_NAME}
+        :param install_profile: The profile to use when building targets to install. Defaults to "release".
         """
         self.build_script_path = build_script_path
         self.project_save_path = project_save_path
+        self.install_profile = install_profile
 
-    def install(self, source_dir: str, header_dir: str, lib_dir: str, exec_dir: str):
+    def install(self, source_dir: str, header_dir: str, lib_dir: str, exec_dir: str) -> DependencyMetadata:
         # Configuration scripts should save the project.
         configure_status = subprocess.run([sys.executable, self.build_script_path], capture_output=True, cwd=source_dir)
         if configure_status.returncode:
@@ -27,5 +31,12 @@ class SBuildrBuilder(DependencyBuilder):
         if not os.path.exists(saved_project):
             G_LOGGER.critical(f"Project was not saved to: {saved_project}. Please ensure this path is correct, and that the build configuration script in {self.build_script_path} is saving the project")
 
-        proj = Project.load(saved_project)
-        proj.install(header_install_path=header_dir, library_install_path=lib_dir, executable_install_path=exec_dir, dry_run=False)
+        project = Project.load(saved_project)
+        project.install(profile_names=[self.install_profile], header_install_path=header_dir, library_install_path=lib_dir, executable_install_path=exec_dir, dry_run=False)
+
+        libraries = {}
+        for name, target in project.libraries.items():
+            lib = target[self.install_profile]
+            libraries[name] = Library(lib.path, ld_dirs=lib.ld_dirs)
+        include_dirs = project.files.include_dirs
+        return DependencyMetadata(libraries, include_dirs)